+++ src/main/java/com/osen/sistema_reservas/core/dashboard/application/dtos/IngresoMensualDTO.java
@@ -1,3 +1,5 @@
 package com.osen.sistema_reservas.core.dashboard.application.dtos;
 
-public record IngresoMensualDTO(String mes, double monto) {}
+import java.math.BigDecimal;
+
+public record IngresoMensualDTO(String mes, BigDecimal monto) {}
