+++ src/main/java/com/osen/sistema_reservas/core/detalle_reserva/application/dtos/DetalleReservaResponse.java
@@ -1,7 +1,5 @@
 package com.osen.sistema_reservas.core.detalle_reserva.application.dtos;
 
-public record DetalleReservaResponse(
-        Long id,
-        Long habitacionId,
-        double precioNoche
-) {}
+import java.math.BigDecimal;
+
+public record DetalleReservaResponse(Long id, Long habitacionId, BigDecimal precioNoche) {}
