+++ src/main/java/com/osen/sistema_reservas/core/hotel/application/dtos/HotelDetalleResponse.java
@@ -3,24 +3,19 @@ package com.osen.sistema_reservas.core.hotel.application.dtos;
 import com.osen.sistema_reservas.core.departamento.application.dtos.DepartamentoResponse;
 import com.osen.sistema_reservas.core.habitacion.application.dtos.HabitacionResponse;
 
+import java.math.BigDecimal;
 import java.util.List;
 
-/**
- * DTO para respuesta con detalle completo de un hotel
- */
 public record HotelDetalleResponse(
         HotelInfo hotel,
         DepartamentoResponse departamento,
         List<HabitacionResponse> habitaciones
 ) {
-    /**
-     * Información básica del hotel
-     */
     public record HotelInfo(
             Long id,
             String nombre,
             String direccion,
-            double precioMinimo,
+            BigDecimal precioMinimo,
             int cantidadHabitaciones
     ) {}
 }
