+++ src/main/java/com/osen/sistema_reservas/core/reserva/application/dtos/ReservaUpdateResponse.java
@@ -1,13 +1,9 @@
 package com.osen.sistema_reservas.core.reserva.application.dtos;
 
-/**
- * DTO para respuesta de actualización de reserva
- */
-public record ReservaUpdateResponse(
-        String message,
-        double nuevoTotal
-) {
-    public static ReservaUpdateResponse of(double nuevoTotal) {
+import java.math.BigDecimal;
+
+public record ReservaUpdateResponse(String message, BigDecimal nuevoTotal) {
+    public static ReservaUpdateResponse of(BigDecimal nuevoTotal) {
         return new ReservaUpdateResponse("Reserva actualizada exitosamente", nuevoTotal);
     }
 }
